+++ src/diagnostics/preambleCheck.ts
@@ -6,22 +6,27 @@
 
 import * as vscode from 'vscode';
 import * as parser from '../parser';
-import { Errors } from './common';
+import { Errors, Warnings } from './common';
 import { analyseSignature } from './signatureCheck';
 import { Quest } from '../language/quest';
 import { LanguageData } from '../language/static/languageData';
 
 /**
  * Analyses the Preamble of a quest.
- * @param context Diagnostic context for the current document.
+ * @param quest The quest that owns target preamble.
  * @param data Language data used for linting.
  */
-export function* analysePreamble(context: Quest, data: LanguageData): Iterable<vscode.Diagnostic> {
-    for (const action of context.preamble.directives) {
-        yield* analyseSignature(context, data, action.line, [action.parameter], false);
+export function* analysePreamble(quest: Quest, data: LanguageData): Iterable<vscode.Diagnostic> {
+    for (const directive of quest.preamble.directives) {
+        yield* analyseSignature(quest, data, directive.line, [directive.parameter], false);
+
+        if (!quest.document.isUntitled && !directive.valueRange.isEmpty &&
+            directive.name === 'Quest' && quest.name !== directive.parameter.value) {
+            yield Warnings.questNameMismatch(directive.valueRange);
+        }
     }
 
-    for (const line of context.preamble.failedParse) {
+    for (const line of quest.preamble.failedParse) {
         yield Errors.undefinedExpression(parser.trimRange(line), 'Preamble');
     }
 }